Fix navigation highlight and move to separate component
